sensor_collector/requirements.txt
sense-hat
psycopg2-binary
python-dotenv
